pyproject.toml
[tool.poetry]
name = "rag-system"
description = ""
version = "0.1.0"
authors = [
    "Camilo <ramirezfandino@gmail.com>"
    ]
package-mode = false
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pydantic = "^2.6.3"
pydantic-settings = "^2.1.0"
bytewax = "^0.18.2"
pika = "^1.3.2"
qdrant-client = "^1.8.0"
unstructured = "^0.12.6"
langchain = "^0.1.13"
numpy = "^1.26.4"
langchain-openai = "^0.1.3"
pymongo = "^4.7.1"
structlog = "^24.1.0"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
openai = "^1.30.1"
selenium = "^4.21.0"
pandas = "^2.2.2"
weaviate-client = "^4.6.3"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"
sentence-transformers = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
